requirements.txt
Django
psycopg2-binary
gunicorn
python-dotenv
requests
whitenoise
Pillow
django-rosetta
django-parler